Extract fun from `LocalMediaView` and move some files.
