+++ libraries/mediaviewer/impl/src/main/kotlin/io/element/android/libraries/mediaviewer/impl/local/image/MediaImageView.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * Please see LICENSE in the repository root for full details.
+ */
+
+package io.element.android.libraries.mediaviewer.impl.local.image
+
+import androidx.compose.foundation.Image
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalInspectionMode
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import io.element.android.libraries.designsystem.utils.CommonDrawables
+import io.element.android.libraries.mediaviewer.api.local.LocalMedia
+import io.element.android.libraries.mediaviewer.impl.local.LocalMediaViewState
+import io.element.android.libraries.ui.strings.CommonStrings
+import me.saket.telephoto.zoomable.coil.ZoomableAsyncImage
+import me.saket.telephoto.zoomable.rememberZoomableImageState
+
+@Composable
+fun MediaImageView(
+    localMediaViewState: LocalMediaViewState,
+    localMedia: LocalMedia?,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    if (LocalInspectionMode.current) {
+        Image(
+            painter = painterResource(id = CommonDrawables.sample_background),
+            modifier = modifier,
+            contentDescription = null,
+        )
+    } else {
+        val zoomableImageState = rememberZoomableImageState(localMediaViewState.zoomableState)
+        localMediaViewState.isReady = zoomableImageState.isImageDisplayed
+        ZoomableAsyncImage(
+            modifier = modifier,
+            state = zoomableImageState,
+            model = localMedia?.uri,
+            contentDescription = stringResource(id = CommonStrings.common_image),
+            contentScale = ContentScale.Fit,
+            onClick = { onClick() }
+        )
+    }
+}
