+++ libraries/mediaviewer/impl/src/main/kotlin/io/element/android/libraries/mediaviewer/impl/local/video/MediaPlayerControllerState.kt
@@ -5,7 +5,7 @@
  * Please see LICENSE in the repository root for full details.
  */
 
-package io.element.android.libraries.mediaviewer.impl.player
+package io.element.android.libraries.mediaviewer.impl.local.video
 
 data class MediaPlayerControllerState(
     val isVisible: Boolean,
